Unification of singleton meta class
