+++ nemo/utils/app_state.py
@@ -15,33 +15,9 @@
 # limitations under the License.
 # =============================================================================
 
-import threading
-
 # Sadly have to import this to avoid circular dependencies.
 import nemo
-
-
-class Singleton(type):
-    """ Implementation of a generic, tread-safe singleton meta-class.
-        Can be used as meta-class, i.e. will create 
-    """
-
-    # List of instances - one per class.
-    __instances = {}
-    # Lock used for accessing the instance.
-    __lock = threading.Lock()
-
-    def __call__(cls, *args, **kwargs):
-        """ Returns singleton instance.A thread safe implementation. """
-        if cls not in cls.__instances:
-            # Enter critical section.
-            with cls.__lock:
-                # Check once again.
-                if cls not in cls.__instances:
-                    # Create a new object instance - one per class.
-                    cls.__instances[cls] = super(Singleton, cls).__call__(*args, **kwargs)
-        # Return the instance.
-        return cls.__instances[cls]
+from nemo.utils.metaclasses import Singleton
 
 
 class AppState(metaclass=Singleton):
