+++ nemo/utils/metaclasses.py
@@ -13,17 +13,30 @@
 # limitations under the License.****
 
 __all__ = [
-    "SingletonMetaClass",
+    "Singleton",
 ]
 
+import threading
 
-class SingletonMetaClass(type):
 
-    _instances = {}
+class Singleton(type):
+    """ Implementation of a generic, tread-safe singleton meta-class.
+        Can be used as meta-class, i.e. will create 
+    """
+
+    # List of instances - one per class.
+    __instances = {}
+    # Lock used for accessing the instance.
+    __lock = threading.Lock()
 
     def __call__(cls, *args, **kwargs):
-
-        if cls not in cls._instances:
-            cls._instances[cls] = super(SingletonMetaClass, cls).__call__(*args, **kwargs)
-
-        return cls._instances[cls]
+        """ Returns singleton instance.A thread safe implementation. """
+        if cls not in cls.__instances:
+            # Enter critical section.
+            with cls.__lock:
+                # Check once again.
+                if cls not in cls.__instances:
+                    # Create a new object instance - one per class.
+                    cls.__instances[cls] = super(Singleton, cls).__call__(*args, **kwargs)
+        # Return the instance.
+        return cls.__instances[cls]
